refactor the logger to be a singleton and use init()
Cleanup the logger to better follow Go best practices.

Refs: #1001
