+++ internal/sso/cache_test.go
@@ -29,6 +29,7 @@ import (
 	"github.com/sirupsen/logrus/hooks/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
+	"github.com/synfinatic/aws-sso-cli/internal/logger"
 )
 
 const (
@@ -264,11 +265,11 @@ func (suite *CacheTestSuite) TestDeleteOldHistory() {
 		c.SSO["Default"].Roles.Accounts[123456789012].Roles["Foo"].Tags)
 
 	// setup logger for tests
-	logger, hook := test.NewNullLogger()
-	logger.SetLevel(logrus.DebugLevel)
-	oldLogger := GetLogger()
-	SetLogger(logger)
-	defer SetLogger(oldLogger)
+	logrusLogger, hook := test.NewNullLogger()
+	logrusLogger.SetLevel(logrus.DebugLevel)
+	oldLog := log
+	log = logger.NewLogger(logrusLogger)
+	defer func() { log = oldLog }()
 
 	// remove one because of HistoryMinutes expires
 	c = suite.setupDeleteOldHistory()
