+++ internal/sso/logger.go
@@ -18,21 +18,10 @@ package sso
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import (
-	"github.com/sirupsen/logrus"
-)
+import "github.com/synfinatic/aws-sso-cli/internal/logger"
 
-var log *logrus.Logger
+var log *logger.Logger
 
-func SetLogger(l *logrus.Logger) {
-	log = l
-}
-
-func GetLogger() *logrus.Logger {
-	return log
-}
-
-// this is configured by cmd/main.go, but we have this here for unit tests
 func init() {
-	log = logrus.New()
+	log = logger.GetLogger()
 }
